Retiradas funções da classe User e movidas para as classes para melhorar o encapsulamento
